+++ build_runner_core/test/generate/custom_generated_dir_test.dart
@@ -21,7 +21,7 @@ void main() {
 
   test('can output files to a custom generated dir', () async {
     var writer = InMemoryRunnerAssetWriter();
-    await testBuilders(
+    final result = await testBuilders(
         [
           applyToRoot(
               TestBuilder(
@@ -33,10 +33,9 @@ void main() {
         outputs: {
           r'$$a|lib/a.txt.copy': 'a',
         },
-        expectedGeneratedDir: customGeneratedDir,
-        writer: writer);
+        expectedGeneratedDir: customGeneratedDir);
     expect(
-        writer.assets[AssetId(
+        result.writer.assets[AssetId(
             'a', '.dart_tool/build/$customGeneratedDir/a/lib/a.txt.copy')],
         isNotNull);
   });